+++ react/components/WordpressPage.tsx
@@ -238,7 +238,8 @@ const WordpressPage: StorefrontFunctionComponent<PageProps> = ({
   try {
     parsedCustomDomains = customDomains ? JSON.parse(customDomains) : null
   } catch (e) {
-    console.error(e)
+    console.error(`${e.name}: ${e.message}`)
+    return null
   }
 
   const customDomain =