+++ react/components/WordpressPost.tsx
@@ -312,7 +312,8 @@ const WordpressPost: StorefrontFunctionComponent<PostProps> = ({
   try {
     parsedCustomDomains = customDomains ? JSON.parse(customDomains) : null
   } catch (e) {
-    console.error(e)
+    console.error(`${e.name}: ${e.message}`)
+    return null
   }
 
   const customDomain =