+++ react/components/WordpressSearchResult.tsx
@@ -61,7 +61,8 @@ const WordpressSearchResult: StorefrontFunctionComponent<SearchProps> = ({
   try {
     parsedCustomDomains = customDomains ? JSON.parse(customDomains) : null
   } catch (e) {
-    console.error(e)
+    console.error(`${e.name}: ${e.message}`)
+    return null
   }
   const customDomain =
     params.customdomainslug && parsedCustomDomains